requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
python-dotenv
pydantic